Updated get/post requests to take a :headers parameter.
By default, the underlying get request will send an  Accept 'application/xml'
header.  If a Rails respond_to handler is used and there is no explicit
handling of xml, the controller will return a "406 Not Acceptable".

This was discovered when we had an action that responded to 'html' and 'atom'.
When /resources.atom was requested the response code was 200, when /resources
was requested, the response code was 406.

This commit allows for the passing of headers in the generic sense (see below).

  get(
    "http://yoursite.com/path/to/resource",
    :headers => {:accept => 'text/html'}
  )